+ Added work on TargetCircle setup and trigger code. + Added work on War to handle combat cycle: increase rounds, check game end, check input boundaries, handle target circle creation/setup and action times. * Restored SetTeamInTurn and GetTeamInTurn in GameNetwork that can be reused a lot in War script. * Made methods that do not use local variables in GameNetwork, static.

+++ Assets/A_Game/Scripts/MinMinUnit.cs
@@ -18,7 +18,7 @@ public class MinMinUnit : NetworkEntity
     public int Defense = 1;
     public int MaxHealth = 5;
 
-    public int EffectScale = 1;
+    public float EffectScale = 1;
 
     public int Effect;
     public Types Type;
